Refactor absolute paths to use @/
As in Expo 50 babel config for absolute paths are not required anymore and now is just setup on `tsconfig` file, the refactor was made to remove the absolute paths resolver from babel.config and refactor to use `@hooks/` to `@/hooks/` for example. the `@/` path now points directly for the `src` source, so is possible to access all folders in the source through `@/` and no need to every time add a new folder in the tsconfig anymore.